requirements.txt
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.23.4
asyncio==3.4.3
pytest==7.1.2
pytest-asyncio==0.19.0
lmdb==1.4.1
numpy==1.23.2